+++ Sources/VCore/Extensions/Foundation/Collection.FirstIndexAndElement.swift
@@ -1,5 +1,5 @@
 //
-//  Array.FirstIndexAndElement.swift
+//  Collection.FirstIndexAndElement.swift
 //  VCore
 //
 //  Created by Vakhtang Kontridze on 01.05.22.
@@ -7,9 +7,9 @@
 
 import Foundation
 
-// MARK: - Array First Index and Element
-extension Array {
-    /// Returns first index and element of `Array` that satisfies the given predicate.
+// MARK: - Collection First Index and Element
+extension Collection {
+    /// Returns first index and element of `Collection` that satisfies the given predicate.
     ///
     /// Used for avoiding accessing both index and element with built-in methods.
     ///
@@ -18,8 +18,8 @@ extension Array {
     ///
     public func firstIndexAndElement(
         where predicate: (Element) throws -> Bool
-    ) rethrows -> (index: Int, element: Element)? {
-        guard let index: Int = try firstIndex(where: predicate) else { return nil }
+    ) rethrows -> (index: Index, element: Element)? {
+        guard let index: Index = try firstIndex(where: predicate) else { return nil }
         let element: Element = self[index]
         
         return (index, element)