+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Architectural Pattern Helpers
 
 Extensions
 
+- `Array.firstIndexAndElement(where:)` is expanded to `Collection`
 - `UIView.removeSubviews` is renamed to `UIView.removeSubviewsFromSuperview`
 
 ### [3.18.4(57)](https://github.com/VakhoKontridze/VCore/releases/tag/3.18.4) — *2022 07 26*